# openapi.yml
openapi: 3.0.0
info:
  title: ITT-Orchestrator API
  description: Secure Enterprise Agent Gateway (SEAG) - Production Control Plane API
  version: 0.1.0
  contact:
    name: ITT-Orchestrator Team
    url: https://github.com/your-org/itt-orchestrator
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token. Obtain from POST /auth/login
  schemas:
    Error:
      type: object
      required:
        - error_id
        - status
        - message
      properties:
        error_id:
          type: string
          format: uuid
          description: Unique error identifier for tracking
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
        details:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Bearer token
        expires_in:
          type: integer
          description: Token expiry in seconds
        token_type:
          type: string
          example: Bearer
...
    ApiRegistryEntry:
      type: object
      required:
        - id
        - name
        - spec_link
      properties:
        id:
          type: string
        name:
          type: string
          description: API name
        spec_link:
          type: string
          description: Link to OpenAPI specification
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Zone:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
          description: Zone name (e.g., Zone 1, Zone 4)
        description:
          type: string
        ips:
          type: array
          items:
            type: string
            format: ipv4
        filters:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Login endpoint
      description: |
        Authenticate user and receive JWT token.
        
        Default credentials for development:
        - Username: admin, Password: any
        - Username: user, Password: any
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registry:
    get:
      operationId: getRegistry
      summary: Get API Registry
      description: Fetches all discovered and registered APIs
      tags:
        - API Registry
      responses:
        '200':
          description: List of registered APIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiRegistryEntry'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /zones:
    get:
      operationId: getZones
      summary: Get Adaptive Gateway Fabric Zones
      tags:
        - Zones
      responses:
        '200':
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zone'

    post:
      operationId: postZone
      summary: Create Zone
      description: Create a new Adaptive Gateway Fabric zone
      tags:
        - Zones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Zone'
      responses:
        '201':
          description: Zone created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
...